+++ src/bonus/main.c
@@ -6,57 +6,11 @@
 /*   By: ttsubo <ttsubo@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/11 19:07:46 by ttsubo            #+#    #+#             */
-/*   Updated: 2025/03/07 11:46:42 by ttsubo           ###   ########.fr       */
+/*   Updated: 2025/03/07 12:02:41 by ttsubo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "bonus.h"
-
-void	exec(char *cmd, char **ev)
-{
-	char	*bin_path;
-	char	**cmds;
-
-	cmds = ft_split(cmd, ' ');
-	if (!cmds)
-		error("exec");
-	bin_path = get_cmd_path(cmds[0], ev);
-	if (!bin_path)
-	{
-		ft_putstr_fd(cmds[0], STDERR_FILENO);
-		ft_putstr_fd(": command not found\n", STDERR_FILENO);
-		exec_free(bin_path, cmds);
-		exit(1);
-	}
-	if (execve(bin_path, cmds, ev) == -1)
-		error(cmds[0]);
-}
-
-void	child_process(char *cmd, char **ev)
-{
-	pid_t	pid;
-	int		fd[2];
-
-	if (pipe(fd) == -1)
-		error("child_process: pipe failed");
-	pid = fork();
-	if (pid == -1)
-		error("child_process: fork failed");
-	if (pid == 0)
-	{
-		close(fd[0]);
-		dup2(fd[1], STDOUT_FILENO);
-		close(fd[1]);
-		exec(cmd, ev);
-	}
-	else
-	{
-		close(fd[1]);
-		dup2(fd[0], STDIN_FILENO);
-		close(fd[0]);
-		waitpid(pid, NULL, 0);
-	}
-}
 
 int	main(int ac, char **av, char **ev)
 {
